Add Verification Test on the reg page Add class NotReadyTests
